backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Cors
google-generativeai
gunicorn